fixed calculate money per day

+++ app/src/main/java/com/example/sampleapp/ui/ProgressViewModel.kt
@@ -23,7 +23,7 @@ class ProgressViewModel(application: Application) : AndroidViewModel(application
     }
 
     private fun calculateMoneyPerDay(smokedPerDay: Int, packPrice: Float, packCount: Int): Float {
-        return ((smokedPerDay / packCount).toFloat() * packPrice)
+        return (smokedPerDay.toFloat() / packCount.toFloat()) * packPrice
     }
 
     fun calculateSavedMoney(): Float? {
